Move signal test util to utils
